Create HomeWork2
Created classes Animal, Cat and Dog, where Cat and Dog are inherited from Animal. Each class has methods run() and swim() that check the restrictions on the actions and display the appropriate messages to the console. The animals array contains Cat and Dog instances, which then loop through the run() and swim() actions. The number of animals created is counted using the animalCount static field.